updated test after namespace and double->float

+++ tests/loading_test.cpp
@@ -4,7 +4,7 @@ int main() {
 
     LPP::Network my_model("saving_test.txt");
 
-    const std::vector<double> classification = my_model.inference(
+    const std::vector<float> classification = my_model.inference(
         {5, 13, -1, 1, 1.4, 0.23, 0, 0, 1, 0}
     );
